Rain component added
