+++ client/src/components/rain/Drop.svelte
