+++ client/src/svgs/CloudRight.svelte
